pyproject.toml
[tool.poetry]
name = "access-logger"
version = "0.1.0"
description = "Read and process nginx log"
authors = ["catbaron0 <catbaron@live.cn>"]
readme = "README.md"
packages = [{include = "access_logger"}]

[tool.poetry.dependencies]
python = "^3.13"
peewee = "^3.17.8"
geoip2 = "^4.8.1"
qqwry-py3 = "^1.2.1"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
